deleteShouldReturnBadRequestWhenDependentId

+++ src/main/java/com/devsuperior/demo/services/CityService.java
@@ -37,10 +37,14 @@ public class CityService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
         try{
             repository.deleteById(id);
-        } catch (DataIntegrityViolationException e) {
-            throw new DatabaseException("Referential integrity error");
+        }
+        catch(DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial.");
         }
     }
 }
